requirements.txt
asgi-ipc==1.3.1
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
BTrees==4.4.1
certifi==2018.1.18
cffi==1.11.5
channels==1.1.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
daphne==1.4.2
DateTime==4.2
defusedxml==0.5.0
Django==1.11.12
django-allauth==0.35.0
django-crispy-forms==1.7.2
django-model-utils==3.1.1
django-webpack-loader==0.6.0
ExtensionClass==4.3.0
gunicorn==19.7.1
h2==3.0.1
haikunator==1.0.1
hpack==3.0.0
hyperframe==5.1.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.19
msgpack-python==0.5.6
oauthlib==2.0.7
PasteDeploy==1.5.2
persistent==4.2.4.2
pkg-resources==0.0.0
posix-ipc==1.0.4
priority==1.3.0
psycopg2==2.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyOpenSSL==17.5.0
python-gettext==3.0
python3-openid==3.1.0
pytz==2018.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
RestrictedPython==3.6.0
service-identity==17.0.0
six==1.11.0
transaction==2.2.1
Twisted==17.9.0
txaio==2.9.0
urllib3==1.22
whitenoise==3.3.1
zc.lockfile==1.2.1
ZConfig==3.2.0
zdaemon==4.2.0
ZEO==5.2.0
zExceptions==4.0
ZODB==5.4.0
ZODB3==3.11.0
zodbpickle==1.0
zope.annotation==4.6.0
zope.browser==2.2.0
zope.browserpage==4.2.0
zope.component==4.4.1
zope.configuration==4.1.0
zope.container==4.2.1
zope.contentprovider==4.1.0
zope.contenttype==4.3.0
zope.deferredimport==4.2.1
zope.deprecation==4.3.0
zope.dottedname==4.2
zope.event==4.3.0
zope.exceptions==4.2.0
zope.filerepresentation==4.2.0
zope.i18n==4.3.1
zope.i18nmessageid==4.1.0
zope.interface==4.4.3
zope.lifecycleevent==4.2.0
zope.location==4.1.0
zope.pagetemplate==4.3.0
zope.processlifetime==2.2.0
zope.proxy==4.3.0
zope.publisher==4.3.2
zope.schema==4.5.0
zope.security==4.2.2
zope.site==4.1.0
zope.size==4.2.0
zope.tal==4.3.1
zope.tales==4.2.0
zope.traversing==4.2.0
zope.viewlet==4.1.0
